don't recheck in -Ycheck
