+++ compiler/src/dotty/tools/dotc/typer/RefChecks.scala
@@ -18,7 +18,6 @@ import config.{ScalaVersion, NoScalaVersion}
 import Decorators._
 import typer.ErrorReporting._
 import DenotTransformers._
-import ValueClasses.{isDerivedValueClass, isCyclic}
 
 object RefChecks {
   import tpd._
@@ -685,42 +684,6 @@ object RefChecks {
         ctx.deprecationWarning(
           symbol.toString + " overrides concrete, non-deprecated symbol(s):" +
             concrOvers.map(_.name.decode).mkString("    ", ", ", ""), tree.pos)
-    }
-  }
-
-  /** Verify classes extending AnyVal meet the requirements */
-  def checkDerivedValueClass(clazz: Symbol, stats: List[Tree])(implicit ctx: Context) = {
-    def checkValueClassMember(stat: Tree) = stat match {
-      case _: ValDef if !stat.symbol.is(ParamAccessor) =>
-        ctx.error(s"value class may not define non-parameter field", stat.pos)
-      case _: DefDef if stat.symbol.isConstructor =>
-        ctx.error(s"value class may not define secondary constructor", stat.pos)
-      case _: MemberDef | _: Import | EmptyTree =>
-      // ok
-      case _ =>
-        ctx.error(s"value class may not contain initialization statements", stat.pos)
-    }
-    if (isDerivedValueClass(clazz)) {
-      if (clazz.is(Trait))
-        ctx.error("Only classes (not traits) are allowed to extend AnyVal", clazz.pos)
-      if (clazz.is(Abstract))
-        ctx.error("`abstract' modifier cannot be used with value classes", clazz.pos)
-      if (!clazz.isStatic)
-        ctx.error(s"value class may not be a ${if (clazz.owner.isTerm) "local class" else "member of another class"}", clazz.pos)
-      if (isCyclic(clazz.asClass))
-        ctx.error("value class cannot wrap itself", clazz.pos)
-      else {
-        val clParamAccessors = clazz.asClass.paramAccessors.filter(_.isTerm)
-        clParamAccessors match {
-          case List(param) =>
-            if (param.is(Mutable))
-              ctx.error("value class parameter must not be a var", param.pos)
-
-          case _ =>
-            ctx.error("value class needs to have exactly one val parameter", clazz.pos)
-        }
-      }
-      stats.foreach(checkValueClassMember)
     }
   }
